@@ -1,9 +1,13 @@
 package main
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"log"
 	"os"
+	"os/signal"
+	"syscall"
 	"time"
 
 	"github.com/percona/percona-xtradb-cluster-operator/cmd/pitr/collector"
@@ -17,18 +21,20 @@ func main() {
 	if len(os.Args) > 1 {
 		command = os.Args[1]
 	}
+	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGTERM, os.Interrupt)
+	defer stop()
 	switch command {
 	case "collect":
-		runCollector()
+		runCollector(ctx)
 	case "recover":
-		runRecoverer()
+		runRecoverer(ctx)
 	default:
 		fmt.Fprintf(os.Stderr, "ERROR: unknown command \"%s\".\nCommands:\n  collect - collect binlogs\n  recover - recover from binlogs\n", command)
 		os.Exit(1)
 	}
 }
 
-func runCollector() {
+func runCollector(ctx context.Context) {
 	config, err := getCollectorConfig()
 	if err != nil {
 		log.Fatalln("ERROR: get config:", err)
@@ -39,26 +45,32 @@ func runCollector() {
 	}
 	log.Println("run binlog collector")
 	for {
-		err := c.Run()
+		err := c.Run(ctx)
 		if err != nil {
 			log.Println("ERROR:", err)
 		}
 
-		time.Sleep(time.Duration(config.CollectSpanSec) * time.Second)
+		t := time.NewTimer(time.Duration(config.CollectSpanSec) * time.Second)
+		select {
+		case <-ctx.Done():
+			log.Fatalln("ERROR:", ctx.Err().Error())
+		case <-t.C:
+			break
+		}
 	}
 }
 
-func runRecoverer() {
+func runRecoverer(ctx context.Context) {
 	config, err := getRecovererConfig()
 	if err != nil {
 		log.Fatalln("ERROR: get recoverer config:", err)
 	}
-	c, err := recoverer.New(config)
+	c, err := recoverer.New(ctx, config)
 	if err != nil {
 		log.Fatalln("ERROR: new recoverer controller:", err)
 	}
 	log.Println("run recover")
-	err = c.Run()
+	err = c.Run(ctx)
 	if err != nil {
 		log.Fatalln("ERROR: recover:", err)
 	}
@@ -67,6 +79,18 @@ func runRecoverer() {
 func getCollectorConfig() (collector.Config, error) {
 	cfg := collector.Config{}
 	err := env.Parse(&cfg)
+	switch cfg.StorageType {
+	case "s3":
+		if err := env.Parse(&cfg.BackupStorageS3); err != nil {
+			return cfg, err
+		}
+	case "azure":
+		if err := env.Parse(&cfg.BackupStorageAzure); err != nil {
+			return cfg, err
+		}
+	default:
+		return cfg, errors.New("unknown STORAGE_TYPE")
+	}
 
 	return cfg, err
 
@@ -77,11 +101,23 @@ func getRecovererConfig() (recoverer.Config, error) {
 	if err := env.Parse(&cfg); err != nil {
 		return cfg, err
 	}
-	if err := env.Parse(&cfg.BackupStorage); err != nil {
-		return cfg, err
-	}
-	if err := env.Parse(&cfg.BinlogStorage); err != nil {
-		return cfg, err
+	switch cfg.StorageType {
+	case "s3":
+		if err := env.Parse(&cfg.BackupStorageS3); err != nil {
+			return cfg, err
+		}
+		if err := env.Parse(&cfg.BinlogStorageS3); err != nil {
+			return cfg, err
+		}
+	case "azure":
+		if err := env.Parse(&cfg.BackupStorageAzure); err != nil {
+			return cfg, err
+		}
+		if err := env.Parse(&cfg.BinlogStorageAzure); err != nil {
+			return cfg, err
+		}
+	default:
+		return cfg, errors.New("unknown STORAGE_TYPE")
 	}
 
 	return cfg, nil
