@@ -64,7 +64,7 @@ func TestGetGTIDFromContent(t *testing.T) {
 	c := []byte(`sometext GTID of the last set 'test_set:1-10'
 	`)
 
-	set, err := getGTIDFromContent(c)
+	set, err := getGTIDFromXtrabackup(c)
 	if err != nil {
 		t.Error("get last gtid set", err.Error())
 	}
