@@ -6,16 +6,11 @@
 package v1
 
 import (
-	"strings"
-
-	"github.com/percona/percona-xtradb-cluster-operator/version"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"sigs.k8s.io/controller-runtime/pkg/scheme"
 )
 
 var (
-	mainSchemeGroupVersion = schema.GroupVersion{Group: "pxc.percona.com", Version: strings.Replace("v"+version.Version, ".", "-", -1)}
-	MainSchemeBuilder      = scheme.Builder{GroupVersion: mainSchemeGroupVersion}
 	// SchemeGroupVersion is group version used to register these objects
 	SchemeGroupVersion = schema.GroupVersion{Group: "pxc.percona.com", Version: "v1"}
 
@@ -25,7 +20,11 @@ var (
 
 func init() {
 	SchemeBuilder.Register(
-		&PerconaXtraDBClusterBackup{}, &PerconaXtraDBClusterBackupList{},
+		&PerconaXtraDBCluster{},
+		&PerconaXtraDBClusterList{},
+		&PerconaXtraDBClusterBackup{},
+		&PerconaXtraDBClusterBackupList{},
+		&PerconaXtraDBClusterRestore{},
+		&PerconaXtraDBClusterRestoreList{},
 	)
-	MainSchemeBuilder.Register(&PerconaXtraDBCluster{}, &PerconaXtraDBClusterList{})
 }
