@@ -1,20 +1,22 @@
+//go:build !ignore_autogenerated
 // +build !ignore_autogenerated
 
-// Code generated by operator-sdk. DO NOT EDIT.
+// Code generated by controller-gen. DO NOT EDIT.
 
 package v1
 
 import (
-	metav1 "github.com/jetstack/cert-manager/pkg/apis/meta/v1"
+	apismetav1 "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
 	corev1 "k8s.io/api/core/v1"
-	runtime "k8s.io/apimachinery/pkg/runtime"
-	intstr "k8s.io/apimachinery/pkg/util/intstr"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *AppStatus) DeepCopyInto(out *AppStatus) {
 	*out = *in
-	return
+	out.ComponentStatus = in.ComponentStatus
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new AppStatus.
@@ -28,9 +30,23 @@ func (in *AppStatus) DeepCopy() *AppStatus {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *BackupStorageAzureSpec) DeepCopyInto(out *BackupStorageAzureSpec) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new BackupStorageAzureSpec.
+func (in *BackupStorageAzureSpec) DeepCopy() *BackupStorageAzureSpec {
+	if in == nil {
+		return nil
+	}
+	out := new(BackupStorageAzureSpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *BackupStorageS3Spec) DeepCopyInto(out *BackupStorageS3Spec) {
 	*out = *in
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new BackupStorageS3Spec.
@@ -46,7 +62,16 @@ func (in *BackupStorageS3Spec) DeepCopy() *BackupStorageS3Spec {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *BackupStorageSpec) DeepCopyInto(out *BackupStorageSpec) {
 	*out = *in
-	out.S3 = in.S3
+	if in.S3 != nil {
+		in, out := &in.S3, &out.S3
+		*out = new(BackupStorageS3Spec)
+		**out = **in
+	}
+	if in.Azure != nil {
+		in, out := &in.Azure, &out.Azure
+		*out = new(BackupStorageAzureSpec)
+		**out = **in
+	}
 	if in.Volume != nil {
 		in, out := &in.Volume, &out.Volume
 		*out = new(VolumeSpec)
@@ -59,11 +84,7 @@ func (in *BackupStorageSpec) DeepCopyInto(out *BackupStorageSpec) {
 			(*out)[key] = val
 		}
 	}
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
 	if in.Affinity != nil {
 		in, out := &in.Affinity, &out.Affinity
 		*out = new(corev1.Affinity)
@@ -105,7 +126,11 @@ func (in *BackupStorageSpec) DeepCopyInto(out *BackupStorageSpec) {
 		*out = new(string)
 		**out = **in
 	}
-	return
+	if in.VerifyTLS != nil {
+		in, out := &in.VerifyTLS, &out.VerifyTLS
+		*out = new(bool)
+		**out = **in
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new BackupStorageSpec.
@@ -122,7 +147,6 @@ func (in *BackupStorageSpec) DeepCopy() *BackupStorageSpec {
 func (in *ClusterCondition) DeepCopyInto(out *ClusterCondition) {
 	*out = *in
 	in.LastTransitionTime.DeepCopyInto(&out.LastTransitionTime)
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ClusterCondition.
@@ -136,13 +160,50 @@ func (in *ClusterCondition) DeepCopy() *ClusterCondition {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ComponentStatus) DeepCopyInto(out *ComponentStatus) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ComponentStatus.
+func (in *ComponentStatus) DeepCopy() *ComponentStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(ComponentStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *HAProxySpec) DeepCopyInto(out *HAProxySpec) {
+	*out = *in
+	in.PodSpec.DeepCopyInto(&out.PodSpec)
+	if in.ReplicasServiceEnabled != nil {
+		in, out := &in.ReplicasServiceEnabled, &out.ReplicasServiceEnabled
+		*out = new(bool)
+		**out = **in
+	}
+	if in.ReplicasLoadBalancerSourceRanges != nil {
+		in, out := &in.ReplicasLoadBalancerSourceRanges, &out.ReplicasLoadBalancerSourceRanges
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new HAProxySpec.
+func (in *HAProxySpec) DeepCopy() *HAProxySpec {
+	if in == nil {
+		return nil
+	}
+	out := new(HAProxySpec)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *LogCollectorSpec) DeepCopyInto(out *LogCollectorSpec) {
 	*out = *in
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
 	if in.ContainerSecurityContext != nil {
 		in, out := &in.ContainerSecurityContext, &out.ContainerSecurityContext
 		*out = new(corev1.SecurityContext)
@@ -153,7 +214,6 @@ func (in *LogCollectorSpec) DeepCopyInto(out *LogCollectorSpec) {
 		*out = new(string)
 		**out = **in
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new LogCollectorSpec.
@@ -174,7 +234,6 @@ func (in *PITR) DeepCopyInto(out *PITR) {
 		*out = new(PXCBackupStatus)
 		(*in).DeepCopyInto(*out)
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PITR.
@@ -190,12 +249,7 @@ func (in *PITR) DeepCopy() *PITR {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PITRSpec) DeepCopyInto(out *PITRSpec) {
 	*out = *in
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
-	return
+	in.Resources.DeepCopyInto(&out.Resources)
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PITRSpec.
@@ -211,11 +265,7 @@ func (in *PITRSpec) DeepCopy() *PITRSpec {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PMMSpec) DeepCopyInto(out *PMMSpec) {
 	*out = *in
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
 	if in.ContainerSecurityContext != nil {
 		in, out := &in.ContainerSecurityContext, &out.ContainerSecurityContext
 		*out = new(corev1.SecurityContext)
@@ -226,7 +276,6 @@ func (in *PMMSpec) DeepCopyInto(out *PMMSpec) {
 		*out = new(string)
 		**out = **in
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PMMSpec.
@@ -242,7 +291,6 @@ func (in *PMMSpec) DeepCopy() *PMMSpec {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PXCBackupSpec) DeepCopyInto(out *PXCBackupSpec) {
 	*out = *in
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PXCBackupSpec.
@@ -271,7 +319,18 @@ func (in *PXCBackupStatus) DeepCopyInto(out *PXCBackupStatus) {
 		*out = new(BackupStorageS3Spec)
 		**out = **in
 	}
-	return
+	if in.Azure != nil {
+		in, out := &in.Azure, &out.Azure
+		*out = new(BackupStorageAzureSpec)
+		**out = **in
+	}
+	if in.Conditions != nil {
+		in, out := &in.Conditions, &out.Conditions
+		*out = make([]metav1.Condition, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PXCBackupStatus.
@@ -320,7 +379,11 @@ func (in *PXCScheduledBackup) DeepCopyInto(out *PXCScheduledBackup) {
 		}
 	}
 	in.PITR.DeepCopyInto(&out.PITR)
-	return
+	if in.BackoffLimit != nil {
+		in, out := &in.BackoffLimit, &out.BackoffLimit
+		*out = new(int32)
+		**out = **in
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PXCScheduledBackup.
@@ -336,7 +399,6 @@ func (in *PXCScheduledBackup) DeepCopy() *PXCScheduledBackup {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PXCScheduledBackupSchedule) DeepCopyInto(out *PXCScheduledBackupSchedule) {
 	*out = *in
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PXCScheduledBackupSchedule.
@@ -357,12 +419,19 @@ func (in *PXCSpec) DeepCopyInto(out *PXCSpec) {
 		*out = new(bool)
 		**out = **in
 	}
+	if in.ReplicationChannels != nil {
+		in, out := &in.ReplicationChannels, &out.ReplicationChannels
+		*out = make([]ReplicationChannel, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+	in.Expose.DeepCopyInto(&out.Expose)
 	if in.PodSpec != nil {
 		in, out := &in.PodSpec, &out.PodSpec
 		*out = new(PodSpec)
 		(*in).DeepCopyInto(*out)
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PXCSpec.
@@ -382,7 +451,6 @@ func (in *PerconaXtraDBCluster) DeepCopyInto(out *PerconaXtraDBCluster) {
 	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
 	in.Spec.DeepCopyInto(&out.Spec)
 	in.Status.DeepCopyInto(&out.Status)
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBCluster.
@@ -410,7 +478,6 @@ func (in *PerconaXtraDBClusterBackup) DeepCopyInto(out *PerconaXtraDBClusterBack
 	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
 	out.Spec = in.Spec
 	in.Status.DeepCopyInto(&out.Status)
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterBackup.
@@ -443,7 +510,6 @@ func (in *PerconaXtraDBClusterBackupList) DeepCopyInto(out *PerconaXtraDBCluster
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterBackupList.
@@ -476,7 +542,6 @@ func (in *PerconaXtraDBClusterList) DeepCopyInto(out *PerconaXtraDBClusterList) 
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterList.
@@ -504,7 +569,6 @@ func (in *PerconaXtraDBClusterRestore) DeepCopyInto(out *PerconaXtraDBClusterRes
 	in.ObjectMeta.DeepCopyInto(&out.ObjectMeta)
 	in.Spec.DeepCopyInto(&out.Spec)
 	in.Status.DeepCopyInto(&out.Status)
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterRestore.
@@ -537,7 +601,6 @@ func (in *PerconaXtraDBClusterRestoreList) DeepCopyInto(out *PerconaXtraDBCluste
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterRestoreList.
@@ -571,7 +634,6 @@ func (in *PerconaXtraDBClusterRestoreSpec) DeepCopyInto(out *PerconaXtraDBCluste
 		*out = new(PITR)
 		(*in).DeepCopyInto(*out)
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterRestoreSpec.
@@ -595,7 +657,6 @@ func (in *PerconaXtraDBClusterRestoreStatus) DeepCopyInto(out *PerconaXtraDBClus
 		in, out := &in.LastScheduled, &out.LastScheduled
 		*out = (*in).DeepCopy()
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterRestoreStatus.
@@ -628,7 +689,7 @@ func (in *PerconaXtraDBClusterSpec) DeepCopyInto(out *PerconaXtraDBClusterSpec) 
 	}
 	if in.HAProxy != nil {
 		in, out := &in.HAProxy, &out.HAProxy
-		*out = new(PodSpec)
+		*out = new(HAProxySpec)
 		(*in).DeepCopyInto(*out)
 	}
 	if in.PMM != nil {
@@ -652,7 +713,16 @@ func (in *PerconaXtraDBClusterSpec) DeepCopyInto(out *PerconaXtraDBClusterSpec) 
 		*out = new(bool)
 		**out = **in
 	}
-	return
+	if in.IgnoreAnnotations != nil {
+		in, out := &in.IgnoreAnnotations, &out.IgnoreAnnotations
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.IgnoreLabels != nil {
+		in, out := &in.IgnoreLabels, &out.IgnoreLabels
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterSpec.
@@ -669,6 +739,11 @@ func (in *PerconaXtraDBClusterSpec) DeepCopy() *PerconaXtraDBClusterSpec {
 func (in *PerconaXtraDBClusterStatus) DeepCopyInto(out *PerconaXtraDBClusterStatus) {
 	*out = *in
 	out.PXC = in.PXC
+	if in.PXCReplication != nil {
+		in, out := &in.PXCReplication, &out.PXCReplication
+		*out = new(ReplicationStatus)
+		(*in).DeepCopyInto(*out)
+	}
 	out.ProxySQL = in.ProxySQL
 	out.HAProxy = in.HAProxy
 	out.Backup = in.Backup
@@ -686,7 +761,6 @@ func (in *PerconaXtraDBClusterStatus) DeepCopyInto(out *PerconaXtraDBClusterStat
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PerconaXtraDBClusterStatus.
@@ -712,7 +786,6 @@ func (in *PodAffinity) DeepCopyInto(out *PodAffinity) {
 		*out = new(corev1.Affinity)
 		(*in).DeepCopyInto(*out)
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PodAffinity.
@@ -738,7 +811,6 @@ func (in *PodDisruptionBudgetSpec) DeepCopyInto(out *PodDisruptionBudgetSpec) {
 		*out = new(intstr.IntOrString)
 		**out = **in
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PodDisruptionBudgetSpec.
@@ -752,44 +824,10 @@ func (in *PodDisruptionBudgetSpec) DeepCopy() *PodDisruptionBudgetSpec {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *PodResources) DeepCopyInto(out *PodResources) {
-	*out = *in
-	if in.Requests != nil {
-		in, out := &in.Requests, &out.Requests
-		*out = new(ResourcesList)
-		**out = **in
-	}
-	if in.Limits != nil {
-		in, out := &in.Limits, &out.Limits
-		*out = new(ResourcesList)
-		**out = **in
-	}
-	return
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PodResources.
-func (in *PodResources) DeepCopy() *PodResources {
-	if in == nil {
-		return nil
-	}
-	out := new(PodResources)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *PodSpec) DeepCopyInto(out *PodSpec) {
 	*out = *in
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
-	if in.SidecarResources != nil {
-		in, out := &in.SidecarResources, &out.SidecarResources
-		*out = new(PodResources)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
+	in.SidecarResources.DeepCopyInto(&out.SidecarResources)
 	if in.VolumeSpec != nil {
 		in, out := &in.VolumeSpec, &out.VolumeSpec
 		*out = new(VolumeSpec)
@@ -855,16 +893,39 @@ func (in *PodSpec) DeepCopyInto(out *PodSpec) {
 			(*out)[key] = val
 		}
 	}
+	if in.ServiceLabels != nil {
+		in, out := &in.ServiceLabels, &out.ServiceLabels
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+	if in.ReplicasServiceAnnotations != nil {
+		in, out := &in.ReplicasServiceAnnotations, &out.ReplicasServiceAnnotations
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+	if in.ReplicasServiceLabels != nil {
+		in, out := &in.ReplicasServiceLabels, &out.ReplicasServiceLabels
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
 	if in.ReadinessInitialDelaySeconds != nil {
 		in, out := &in.ReadinessInitialDelaySeconds, &out.ReadinessInitialDelaySeconds
 		*out = new(int32)
 		**out = **in
 	}
+	in.ReadinessProbes.DeepCopyInto(&out.ReadinessProbes)
 	if in.LivenessInitialDelaySeconds != nil {
 		in, out := &in.LivenessInitialDelaySeconds, &out.LivenessInitialDelaySeconds
 		*out = new(int32)
 		**out = **in
 	}
+	in.LivenessProbes.DeepCopyInto(&out.LivenessProbes)
 	if in.PodSecurityContext != nil {
 		in, out := &in.PodSecurityContext, &out.PodSecurityContext
 		*out = new(corev1.PodSecurityContext)
@@ -882,12 +943,25 @@ func (in *PodSpec) DeepCopyInto(out *PodSpec) {
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
+	if in.SidecarVolumes != nil {
+		in, out := &in.SidecarVolumes, &out.SidecarVolumes
+		*out = make([]corev1.Volume, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
+	if in.SidecarPVCs != nil {
+		in, out := &in.SidecarPVCs, &out.SidecarPVCs
+		*out = make([]corev1.PersistentVolumeClaim, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
+	}
 	if in.RuntimeClassName != nil {
 		in, out := &in.RuntimeClassName, &out.RuntimeClassName
 		*out = new(string)
 		**out = **in
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new PodSpec.
@@ -901,17 +975,119 @@ func (in *PodSpec) DeepCopy() *PodSpec {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ResourcesList) DeepCopyInto(out *ResourcesList) {
+func (in *ReplicationChannel) DeepCopyInto(out *ReplicationChannel) {
 	*out = *in
-	return
+	if in.SourcesList != nil {
+		in, out := &in.SourcesList, &out.SourcesList
+		*out = make([]ReplicationSource, len(*in))
+		copy(*out, *in)
+	}
+	if in.Config != nil {
+		in, out := &in.Config, &out.Config
+		*out = new(ReplicationChannelConfig)
+		**out = **in
+	}
 }
 
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ResourcesList.
-func (in *ResourcesList) DeepCopy() *ResourcesList {
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ReplicationChannel.
+func (in *ReplicationChannel) DeepCopy() *ReplicationChannel {
 	if in == nil {
 		return nil
 	}
-	out := new(ResourcesList)
+	out := new(ReplicationChannel)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ReplicationChannelConfig) DeepCopyInto(out *ReplicationChannelConfig) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ReplicationChannelConfig.
+func (in *ReplicationChannelConfig) DeepCopy() *ReplicationChannelConfig {
+	if in == nil {
+		return nil
+	}
+	out := new(ReplicationChannelConfig)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ReplicationChannelStatus) DeepCopyInto(out *ReplicationChannelStatus) {
+	*out = *in
+	out.ReplicationChannelConfig = in.ReplicationChannelConfig
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ReplicationChannelStatus.
+func (in *ReplicationChannelStatus) DeepCopy() *ReplicationChannelStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(ReplicationChannelStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ReplicationSource) DeepCopyInto(out *ReplicationSource) {
+	*out = *in
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ReplicationSource.
+func (in *ReplicationSource) DeepCopy() *ReplicationSource {
+	if in == nil {
+		return nil
+	}
+	out := new(ReplicationSource)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ReplicationStatus) DeepCopyInto(out *ReplicationStatus) {
+	*out = *in
+	if in.Channels != nil {
+		in, out := &in.Channels, &out.Channels
+		*out = make([]ReplicationChannelStatus, len(*in))
+		copy(*out, *in)
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ReplicationStatus.
+func (in *ReplicationStatus) DeepCopy() *ReplicationStatus {
+	if in == nil {
+		return nil
+	}
+	out := new(ReplicationStatus)
+	in.DeepCopyInto(out)
+	return out
+}
+
+// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
+func (in *ServiceExpose) DeepCopyInto(out *ServiceExpose) {
+	*out = *in
+	if in.LoadBalancerSourceRanges != nil {
+		in, out := &in.LoadBalancerSourceRanges, &out.LoadBalancerSourceRanges
+		*out = make([]string, len(*in))
+		copy(*out, *in)
+	}
+	if in.Annotations != nil {
+		in, out := &in.Annotations, &out.Annotations
+		*out = make(map[string]string, len(*in))
+		for key, val := range *in {
+			(*out)[key] = val
+		}
+	}
+}
+
+// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ServiceExpose.
+func (in *ServiceExpose) DeepCopy() *ServiceExpose {
+	if in == nil {
+		return nil
+	}
+	out := new(ServiceExpose)
 	in.DeepCopyInto(out)
 	return out
 }
@@ -926,10 +1102,9 @@ func (in *TLSSpec) DeepCopyInto(out *TLSSpec) {
 	}
 	if in.IssuerConf != nil {
 		in, out := &in.IssuerConf, &out.IssuerConf
-		*out = new(metav1.ObjectReference)
+		*out = new(apismetav1.ObjectReference)
 		**out = **in
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new TLSSpec.
@@ -945,7 +1120,6 @@ func (in *TLSSpec) DeepCopy() *TLSSpec {
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *UpgradeOptions) DeepCopyInto(out *UpgradeOptions) {
 	*out = *in
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new UpgradeOptions.
@@ -975,7 +1149,6 @@ func (in *Volume) DeepCopyInto(out *Volume) {
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new Volume.
@@ -1006,7 +1179,6 @@ func (in *VolumeSpec) DeepCopyInto(out *VolumeSpec) {
 		*out = new(corev1.PersistentVolumeClaimSpec)
 		(*in).DeepCopyInto(*out)
 	}
-	return
 }
 
 // DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new VolumeSpec.
