@@ -6,74 +6,89 @@ import (
 	"fmt"
 	"math/big"
 	mrand "math/rand"
+	"strings"
 	"time"
 
-	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	k8serror "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
+
+	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
+	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/users"
 )
 
+const internalSecretsPrefix = "internal-"
+
 func (r *ReconcilePerconaXtraDBCluster) reconcileUsersSecret(cr *api.PerconaXtraDBCluster) error {
-	secretObj := corev1.Secret{}
+	logger := r.logger(cr.Name, cr.Namespace)
+
+	secretObj := new(corev1.Secret)
 	err := r.client.Get(context.TODO(),
 		types.NamespacedName{
 			Namespace: cr.Namespace,
 			Name:      cr.Spec.SecretsName,
 		},
-		&secretObj,
+		secretObj,
 	)
 	if err == nil {
+		if err := validatePasswords(secretObj); err != nil {
+			return errors.Wrap(err, "validate passwords")
+		}
+		isChanged, err := setUserSecretDefaults(secretObj)
+		if err != nil {
+			return errors.Wrap(err, "set user secret defaults")
+		}
+		if isChanged {
+			err := r.client.Update(context.TODO(), secretObj)
+			if err == nil {
+				logger.Info(fmt.Sprintf("User secrets updated: %s", cr.Spec.SecretsName))
+			}
+			return err
+		}
 		return nil
 	} else if !k8serror.IsNotFound(err) {
 		return errors.Wrap(err, "get secret")
 	}
 
-	data := make(map[string][]byte)
-	data["root"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create root users password")
-	}
-	data["xtrabackup"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create xtrabackup users password")
-	}
-	data["monitor"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create monitor users password")
-	}
-	data["clustercheck"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create clustercheck users password")
-	}
-	data["proxyadmin"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create proxyadmin users password")
-	}
-	data["operator"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "create operator users password")
-	}
-	data["replication"], err = generatePass()
-	if err != nil {
-		return errors.Wrap(err, "generate replication password")
-	}
-
-	secretObj = corev1.Secret{
+	secretObj = &corev1.Secret{
 		ObjectMeta: metav1.ObjectMeta{
 			Name:      cr.Spec.SecretsName,
 			Namespace: cr.Namespace,
 		},
-		Data: data,
 		Type: corev1.SecretTypeOpaque,
 	}
-	err = r.client.Create(context.TODO(), &secretObj)
+
+	if _, err = setUserSecretDefaults(secretObj); err != nil {
+		return errors.Wrap(err, "set user secret defaults")
+	}
+
+	err = r.client.Create(context.TODO(), secretObj)
 	if err != nil {
 		return fmt.Errorf("create Users secret: %v", err)
 	}
+
+	logger.Info(fmt.Sprintf("Created user secrets: %s", cr.Spec.SecretsName))
 	return nil
+}
+
+func setUserSecretDefaults(secret *corev1.Secret) (isChanged bool, err error) {
+	if secret.Data == nil {
+		secret.Data = make(map[string][]byte)
+	}
+	users := []string{"root", "xtrabackup", "monitor", "clustercheck", "proxyadmin", "operator", "replication"}
+	for _, user := range users {
+		if pass, ok := secret.Data[user]; !ok || len(pass) == 0 {
+			secret.Data[user], err = generatePass()
+			if err != nil {
+				return false, errors.Wrapf(err, "create %s users password", user)
+			}
+
+			isChanged = true
+		}
+	}
+	return
 }
 
 const (
@@ -81,10 +96,11 @@ const (
 	passwordMinLen = 16
 	passSymbols    = "ABCDEFGHIJKLMNOPQRSTUVWXYZ" +
 		"abcdefghijklmnopqrstuvwxyz" +
-		"0123456789"
+		"0123456789" +
+		"!#$%&()*+,-.<=>?@[]^_{}~"
 )
 
-//generatePass generate random password
+// generatePass generates a random password
 func generatePass() ([]byte, error) {
 	mrand.Seed(time.Now().UnixNano())
 	ln := mrand.Intn(passwordMaxLen-passwordMinLen) + passwordMinLen
@@ -98,4 +114,19 @@ func generatePass() ([]byte, error) {
 	}
 
 	return b, nil
+}
+
+func validatePasswords(secret *corev1.Secret) error {
+	for user, pass := range secret.Data {
+		switch user {
+		case users.ProxyAdmin:
+			if strings.ContainsAny(string(pass), ";:") {
+				return errors.New("invalid proxyadmin password, don't use ';' or ':'")
+			}
+		default:
+			continue
+		}
+	}
+
+	return nil
 }
