@@ -3,36 +3,37 @@ package pxc
 import (
 	"context"
 	"fmt"
+	"os"
 	"strings"
 	"sync/atomic"
 	"time"
 
-	apiv1 "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
-	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/queries"
 	"github.com/pkg/errors"
 	"github.com/robfig/cron/v3"
 	corev1 "k8s.io/api/core/v1"
 	k8serrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
+	k8sretry "k8s.io/client-go/util/retry"
 	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	apiv1 "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
+	"github.com/percona/percona-xtradb-cluster-operator/pkg/k8s"
+	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/queries"
+	"github.com/percona/percona-xtradb-cluster-operator/version"
 )
 
-const never = "never"
-const disabled = "disabled"
+var versionNotReadyErr = errors.New("not ready to fetch version")
 
 func (r *ReconcilePerconaXtraDBCluster) deleteEnsureVersion(jobName string) {
 	r.crons.crons.Remove(cron.EntryID(r.crons.ensureVersionJobs[jobName].ID))
 	delete(r.crons.ensureVersionJobs, jobName)
 }
 
-func (r *ReconcilePerconaXtraDBCluster) sheduleEnsurePXCVersion(cr *apiv1.PerconaXtraDBCluster, vs VersionService) error {
+func (r *ReconcilePerconaXtraDBCluster) scheduleEnsurePXCVersion(cr *apiv1.PerconaXtraDBCluster, vs VersionService) error {
 	jn := jobName(cr)
 	schedule, ok := r.crons.ensureVersionJobs[jn]
-	if cr.Spec.UpdateStrategy != apiv1.SmartUpdateStatefulSetStrategyType ||
-		cr.Spec.UpgradeOptions.Schedule == "" ||
-		strings.ToLower(cr.Spec.UpgradeOptions.Apply) == never ||
-		strings.ToLower(cr.Spec.UpgradeOptions.Apply) == disabled {
+	if cr.Spec.UpgradeOptions.Schedule == "" || !(versionUpgradeEnabled(cr) || telemetryEnabled()) {
 		if ok {
 			r.deleteEnsureVersion(jn)
 		}
@@ -58,7 +59,7 @@ func (r *ReconcilePerconaXtraDBCluster) sheduleEnsurePXCVersion(cr *apiv1.Percon
 	l := r.lockers.LoadOrCreate(nn.String())
 
 	logger.Info("add new job", "schedule", cr.Spec.UpgradeOptions.Schedule)
-	id, err := r.crons.crons.AddFunc(cr.Spec.UpgradeOptions.Schedule, func() {
+	id, err := r.crons.AddFuncWithSeconds(cr.Spec.UpgradeOptions.Schedule, func() {
 		l.statusMutex.Lock()
 		defer l.statusMutex.Unlock()
 
@@ -84,7 +85,7 @@ func (r *ReconcilePerconaXtraDBCluster) sheduleEnsurePXCVersion(cr *apiv1.Percon
 			return
 		}
 
-		_, err = localCr.CheckNSetDefaults(r.serverVersion, r.logger(cr.Name, cr.Namespace))
+		err = localCr.CheckNSetDefaults(r.serverVersion, r.logger(cr.Name, cr.Namespace))
 		if err != nil {
 			logger.Error(err, "failed to set defaults for CR")
 			return
@@ -119,10 +120,7 @@ func jobName(cr *apiv1.PerconaXtraDBCluster) string {
 }
 
 func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDBCluster, vs VersionService) error {
-	if cr.Spec.UpdateStrategy != apiv1.SmartUpdateStatefulSetStrategyType ||
-		cr.Spec.UpgradeOptions.Schedule == "" ||
-		strings.ToLower(cr.Spec.UpgradeOptions.Apply) == never ||
-		strings.ToLower(cr.Spec.UpgradeOptions.Apply) == disabled {
+	if !(versionUpgradeEnabled(cr) || telemetryEnabled()) {
 		return nil
 	}
 
@@ -130,7 +128,12 @@ func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDB
 		return errors.New("cluster is not ready")
 	}
 
-	newVersion, err := vs.GetExactVersion(cr, cr.Spec.UpgradeOptions.VersionServiceEndpoint, versionMeta{
+	watchNs, err := k8s.GetWatchNamespace()
+	if err != nil {
+		return errors.Wrap(err, "get WATCH_NAMESPACE env variable")
+	}
+
+	vm := versionMeta{
 		Apply:               cr.Spec.UpgradeOptions.Apply,
 		Platform:            string(cr.Spec.Platform),
 		KubeVersion:         r.serverVersion.Info.GitVersion,
@@ -141,12 +144,27 @@ func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDB
 		BackupVersion:       cr.Status.Backup.Version,
 		LogCollectorVersion: cr.Status.LogCollector.Version,
 		CRUID:               string(cr.GetUID()),
-	})
+		ClusterWideEnabled:  watchNs == "",
+	}
+	logger := r.logger(cr.Name, cr.Namespace)
+
+	if telemetryEnabled() && (!versionUpgradeEnabled(cr) || cr.Spec.UpgradeOptions.VersionServiceEndpoint != apiv1.GetDefaultVersionServiceEndpoint()) {
+		_, err := vs.GetExactVersion(cr, apiv1.GetDefaultVersionServiceEndpoint(), vm)
+		if err != nil {
+			logger.Error(err, "failed to send telemetry to "+apiv1.GetDefaultVersionServiceEndpoint())
+		}
+	}
+
+	if !versionUpgradeEnabled(cr) {
+		return nil
+	}
+
+	newVersion, err := vs.GetExactVersion(cr, cr.Spec.UpgradeOptions.VersionServiceEndpoint, vm)
 	if err != nil {
 		return errors.Wrap(err, "failed to check version")
 	}
 
-	logger := r.logger(cr.Name, cr.Namespace)
+	patch := client.MergeFrom(cr.DeepCopy())
 
 	if cr.Spec.PXC != nil && cr.Spec.PXC.Image != newVersion.PXCImage {
 		if cr.Status.PXC.Version == "" {
@@ -202,16 +220,9 @@ func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDB
 		cr.Spec.LogCollector.Image = newVersion.LogCollectorImage
 	}
 
-	err = r.client.Update(context.Background(), cr)
+	err = r.client.Patch(context.Background(), cr.DeepCopy(), patch)
 	if err != nil {
 		return errors.Wrap(err, "failed to update CR")
-	}
-
-	time.Sleep(1 * time.Second) // based on experiments operator just need it.
-
-	err = r.client.Get(context.TODO(), types.NamespacedName{Name: cr.Name, Namespace: cr.Namespace}, cr)
-	if err != nil {
-		logger.Error(err, "failed to get CR")
 	}
 
 	cr.Status.ProxySQL.Version = newVersion.ProxySqlVersion
@@ -222,7 +233,18 @@ func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDB
 	cr.Status.PXC.Image = newVersion.PXCImage
 	cr.Status.LogCollector.Version = newVersion.LogCollectorVersion
 
-	err = r.client.Status().Update(context.Background(), cr)
+	err = k8sretry.RetryOnConflict(k8sretry.DefaultRetry, func() error {
+		localCr := &apiv1.PerconaXtraDBCluster{}
+
+		err := r.client.Get(context.TODO(), types.NamespacedName{Name: cr.Name, Namespace: cr.Namespace}, localCr)
+		if err != nil {
+			return err
+		}
+
+		localCr.Status = cr.Status
+
+		return r.client.Status().Update(context.TODO(), localCr)
+	})
 	if err != nil {
 		return errors.Wrap(err, "failed to update CR status")
 	}
@@ -232,25 +254,25 @@ func (r *ReconcilePerconaXtraDBCluster) ensurePXCVersion(cr *apiv1.PerconaXtraDB
 	return nil
 }
 
-func (r *ReconcilePerconaXtraDBCluster) fetchVersionFromPXC(cr *apiv1.PerconaXtraDBCluster, sfs apiv1.StatefulApp) error {
-	if cr.Status.PXC.Status != apiv1.AppStateReady {
-		return nil
+func (r *ReconcilePerconaXtraDBCluster) mysqlVersion(cr *apiv1.PerconaXtraDBCluster, sfs apiv1.StatefulApp) (string, error) {
+	if cr.Status.PXC.Ready < 1 {
+		return "", versionNotReadyErr
 	}
 
 	if cr.Status.ObservedGeneration != cr.ObjectMeta.Generation {
-		return nil
+		return "", versionNotReadyErr
 	}
 
 	if cr.Status.PXC.Image == cr.Spec.PXC.Image {
-		return nil
+		return "", versionNotReadyErr
 	}
 
 	upgradeInProgress, err := r.upgradeInProgress(cr, "pxc")
 	if err != nil {
-		return errors.Wrap(err, "check pxc upgrade progress")
+		return "", errors.Wrap(err, "check pxc upgrade progress")
 	}
 	if upgradeInProgress {
-		return nil
+		return "", versionNotReadyErr
 	}
 
 	list := corev1.PodList{}
@@ -261,7 +283,7 @@ func (r *ReconcilePerconaXtraDBCluster) fetchVersionFromPXC(cr *apiv1.PerconaXtr
 			LabelSelector: labels.SelectorFromSet(sfs.Labels()),
 		},
 	); err != nil {
-		return errors.Wrap(err, "get pod list")
+		return "", errors.Wrap(err, "get pod list")
 	}
 
 	user := "root"
@@ -275,7 +297,11 @@ func (r *ReconcilePerconaXtraDBCluster) fetchVersionFromPXC(cr *apiv1.PerconaXtr
 	logger := r.logger(cr.Name, cr.Namespace)
 
 	for _, pod := range list.Items {
-		database, err := queries.New(r.client, cr.Namespace, secrets, user, pod.Name+"."+cr.Name+"-pxc."+cr.Namespace, port)
+		if !isPodReady(pod) {
+			continue
+		}
+
+		database, err := queries.New(r.client, cr.Namespace, secrets, user, pod.Name+"."+cr.Name+"-pxc."+cr.Namespace, port, cr.Spec.PXC.ReadinessProbes.TimeoutSeconds)
 		if err != nil {
 			logger.Error(err, "failed to create db instance")
 			continue
@@ -289,16 +315,79 @@ func (r *ReconcilePerconaXtraDBCluster) fetchVersionFromPXC(cr *apiv1.PerconaXtr
 			continue
 		}
 
-		logger.Info("update PXC version (fetched from db)", "new version", version)
-		cr.Status.PXC.Version = version
-		cr.Status.PXC.Image = cr.Spec.PXC.Image
-		err = r.client.Status().Update(context.Background(), cr)
-		if err != nil {
-			return errors.Wrap(err, "failed to update CR")
-		}
+		return version, nil
+	}
 
+	return "", errors.New("failed to reach any pod")
+}
+
+func (r *ReconcilePerconaXtraDBCluster) fetchVersionFromPXC(cr *apiv1.PerconaXtraDBCluster, sfs apiv1.StatefulApp) error {
+	logger := r.logger(cr.Name, cr.Namespace)
+
+	if cr.Status.PXC.Status != apiv1.AppStateReady {
 		return nil
 	}
 
-	return errors.New("failed to reach any pod")
+	version, err := r.mysqlVersion(cr, sfs)
+	if err != nil {
+		if errors.Is(err, versionNotReadyErr) {
+			return nil
+		}
+
+		return err
+	}
+
+	cr.Status.PXC.Version = version
+	cr.Status.PXC.Image = cr.Spec.PXC.Image
+
+	logger.Info("update PXC version (fetched from db)", "new version", version)
+	err = k8sretry.RetryOnConflict(k8sretry.DefaultRetry, func() error {
+		localCr := &apiv1.PerconaXtraDBCluster{}
+
+		err := r.client.Get(context.TODO(), types.NamespacedName{Name: cr.Name, Namespace: cr.Namespace}, localCr)
+		if err != nil {
+			return err
+		}
+
+		localCr.Status = cr.Status
+
+		return r.client.Status().Update(context.TODO(), localCr)
+	})
+	if err != nil {
+		return errors.Wrap(err, "failed to update CR")
+	}
+	return nil
+}
+
+func telemetryEnabled() bool {
+	value, ok := os.LookupEnv("DISABLE_TELEMETRY")
+	if ok {
+		return value != "true"
+	}
+	return true
+}
+
+func versionUpgradeEnabled(cr *apiv1.PerconaXtraDBCluster) bool {
+	return strings.ToLower(cr.Spec.UpgradeOptions.Apply) != apiv1.UpgradeStrategyNever &&
+		strings.ToLower(cr.Spec.UpgradeOptions.Apply) != apiv1.UpgradeStrategyDisabled
+}
+
+// setCRVersion sets operator version of PerconaXtraDBCluster.
+// The new (semver-matching) version is determined by the CR's crVersion field.
+// If the crVersion is an empty string, it sets the current operator version.
+func (r *ReconcilePerconaXtraDBCluster) setCRVersion(ctx context.Context, cr *apiv1.PerconaXtraDBCluster) error {
+	if len(cr.Spec.CRVersion) > 0 {
+		return nil
+	}
+
+	orig := cr.DeepCopy()
+	cr.Spec.CRVersion = version.Version
+
+	if err := r.client.Patch(ctx, cr, client.MergeFrom(orig)); err != nil {
+		return errors.Wrap(err, "patch CR")
+	}
+
+	r.logger(cr.Name, cr.Namespace).Info("Set CR version", "version", cr.Spec.CRVersion)
+
+	return nil
 }
