@@ -2,6 +2,8 @@ package pxcbackup
 
 import (
 	"context"
+	"fmt"
+	"net/url"
 	"os"
 	"reflect"
 	"strconv"
@@ -9,13 +11,11 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
 	"github.com/go-logr/logr"
 	"github.com/go-logr/zapr"
 	"github.com/minio/minio-go/v7"
 	"github.com/minio/minio-go/v7/pkg/credentials"
-	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
-	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/backup"
-	"github.com/percona/percona-xtradb-cluster-operator/version"
 	"github.com/pkg/errors"
 	"go.uber.org/zap"
 	batchv1 "k8s.io/api/batch/v1"
@@ -28,10 +28,15 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
-	"sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
+
+	"github.com/percona/percona-xtradb-cluster-operator/clientcmd"
+	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
+	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/app/deployment"
+	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/backup"
+	"github.com/percona/percona-xtradb-cluster-operator/version"
 )
 
 // Add creates a new PerconaXtraDBClusterBackup Controller and adds it to the Manager. The Manager will set fields on the Controller
@@ -69,10 +74,16 @@ func newReconciler(mgr manager.Manager) (reconcile.Reconciler, error) {
 		return nil, errors.Wrap(err, "failed to create logger")
 	}
 
+	cli, err := clientcmd.NewClient()
+	if err != nil {
+		return nil, errors.Wrap(err, "create clientcmd")
+	}
+
 	return &ReconcilePerconaXtraDBClusterBackup{
 		client:              mgr.GetClient(),
 		scheme:              mgr.GetScheme(),
 		serverVersion:       sv,
+		clientcmd:           cli,
 		chLimit:             make(chan struct{}, limit),
 		bcpDeleteInProgress: new(sync.Map),
 		log:                 zapr.NewLogger(zapLog),
@@ -106,14 +117,14 @@ type ReconcilePerconaXtraDBClusterBackup struct {
 	scheme *runtime.Scheme
 
 	serverVersion       *version.ServerVersion
+	clientcmd           *clientcmd.Client
 	chLimit             chan struct{}
 	bcpDeleteInProgress *sync.Map
 	log                 logr.Logger
 }
 
 func (r *ReconcilePerconaXtraDBClusterBackup) logger(name, namespace string) logr.Logger {
-	return log.NewDelegatingLogger(r.log).WithName("perconaxtradbclusterbackup").
-		WithValues("backup", name, "namespace", namespace)
+	return r.log.WithName("perconaxtradbclusterbackup").WithValues("backup", name, "namespace", namespace)
 }
 
 // Reconcile reads that state of the cluster for a PerconaXtraDBClusterBackup object and makes changes based on the state read
@@ -121,7 +132,7 @@ func (r *ReconcilePerconaXtraDBClusterBackup) logger(name, namespace string) log
 // Note:
 // The Controller will requeue the Request to be processed again if the returned error is non-nil or
 // Result.Requeue is true, otherwise upon completion it will remove the work from the queue.
-func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Request) (reconcile.Result, error) {
+func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(ctx context.Context, request reconcile.Request) (reconcile.Result, error) {
 	logger := r.logger(request.Name, request.Namespace)
 
 	rr := reconcile.Result{
@@ -142,7 +153,7 @@ func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Reques
 		return reconcile.Result{}, err
 	}
 
-	err = r.tryRunS3BackupFinalizerJob(cr)
+	err = r.tryRunBackupFinalizers(ctx, cr)
 	if err != nil {
 		return reconcile.Result{}, errors.Wrap(err, "failed to run finalizers")
 	}
@@ -159,13 +170,13 @@ func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Reques
 		return rr, nil
 	}
 
-	cluster, err := r.getClusterConfig(cr)
+	cluster, err := r.getCluster(cr)
 	if err != nil {
 		logger.Error(err, "invalid backup cluster")
 		return rr, nil
 	}
 
-	_, err = cluster.CheckNSetDefaults(r.serverVersion, logger)
+	err = cluster.CheckNSetDefaults(r.serverVersion, logger)
 	if err != nil {
 		return rr, errors.Wrap(err, "wrong PXC options")
 	}
@@ -178,27 +189,33 @@ func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Reques
 		return rr, errors.Wrap(err, "failed to run backup")
 	}
 
-	bcpStorage, ok := cluster.Spec.Backup.Storages[cr.Spec.StorageName]
+	storage, ok := cluster.Spec.Backup.Storages[cr.Spec.StorageName]
 	if !ok {
-		return rr, errors.Errorf("bcpStorage %s doesn't exist", cr.Spec.StorageName)
+		return rr, errors.Errorf("storage %s doesn't exist", cr.Spec.StorageName)
+	}
+	if cr.Status.S3 == nil || cr.Status.Azure == nil {
+		cr.Status.S3 = storage.S3
+		cr.Status.Azure = storage.Azure
+		cr.Status.StorageType = storage.Type
+		cr.Status.Image = cluster.Spec.Backup.Image
+		cr.Status.SSLSecretName = cluster.Spec.PXC.SSLSecretName
+		cr.Status.SSLInternalSecretName = cluster.Spec.PXC.SSLInternalSecretName
+		cr.Status.VaultSecretName = cluster.Spec.PXC.VaultSecretName
 	}
 
 	bcp := backup.New(cluster)
 	job := bcp.Job(cr, cluster)
-	job.Spec, err = bcp.JobSpec(cr.Spec, cluster.Spec, job)
+	job.Spec, err = bcp.JobSpec(cr.Spec, cluster, job)
 	if err != nil {
 		return rr, errors.Wrap(err, "can't create job spec")
 	}
 
-	var destination string
-	var s3status *api.BackupStorageS3Spec
-
-	switch bcpStorage.Type {
+	switch storage.Type {
 	case api.BackupStorageFilesystem:
 		pvc := backup.NewPVC(cr)
-		pvc.Spec = *bcpStorage.Volume.PersistentVolumeClaim
+		pvc.Spec = *storage.Volume.PersistentVolumeClaim
 
-		destination = "pvc/" + pvc.Name
+		cr.Status.Destination = "pvc/" + pvc.Name
 
 		// Set PerconaXtraDBClusterBackup instance as the owner and controller
 		if err := setControllerReference(cr, pvc, r.scheme); err != nil {
@@ -217,22 +234,32 @@ func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Reques
 			return rr, errors.Wrap(err, "get backup pvc")
 		}
 
-		err := bcp.SetStoragePVC(&job.Spec, cluster, pvc.Name)
+		err := backup.SetStoragePVC(&job.Spec, cr, pvc.Name)
 		if err != nil {
 			return rr, errors.Wrap(err, "set storage FS")
 		}
 	case api.BackupStorageS3:
-		destination = bcpStorage.S3.Bucket + "/" + cr.Spec.PXCCluster + "-" + cr.CreationTimestamp.Time.Format("2006-01-02-15:04:05") + "-full"
-		if !strings.HasPrefix(bcpStorage.S3.Bucket, "s3://") {
-			destination = "s3://" + destination
+		if storage.S3 == nil {
+			return rr, errors.New("s3 storage is not specified")
+		}
+		cr.Status.Destination = storage.S3.Bucket + "/" + cr.Spec.PXCCluster + "-" + cr.CreationTimestamp.Time.Format("2006-01-02-15:04:05") + "-full"
+		if !strings.HasPrefix(storage.S3.Bucket, "s3://") {
+			cr.Status.Destination = "s3://" + cr.Status.Destination
 		}
 
-		err := bcp.SetStorageS3(&job.Spec, cluster, bcpStorage.S3, destination)
+		err := backup.SetStorageS3(&job.Spec, cr)
 		if err != nil {
 			return rr, errors.Wrap(err, "set storage FS")
 		}
-
-		s3status = &bcpStorage.S3
+	case api.BackupStorageAzure:
+		if storage.Azure == nil {
+			return rr, errors.New("azure storage is not specified")
+		}
+		cr.Status.Destination = storage.Azure.ContainerPath + "/" + cr.Spec.PXCCluster + "-" + cr.CreationTimestamp.Time.Format("2006-01-02-15:04:05") + "-full"
+		err := backup.SetStorageAzure(&job.Spec, cr)
+		if err != nil {
+			return rr, errors.Wrap(err, "set storage FS for Azure")
+		}
 	}
 
 	// Set PerconaXtraDBClusterBackup instance as the owner and controller
@@ -247,46 +274,14 @@ func (r *ReconcilePerconaXtraDBClusterBackup) Reconcile(request reconcile.Reques
 		logger.Info("Created a new backup job", "Namespace", job.Namespace, "Name", job.Name)
 	}
 
-	err = r.updateJobStatus(cr, job, destination, cr.Spec.StorageName, s3status)
+	err = r.updateJobStatus(cr, job, cr.Spec.StorageName, storage, cluster)
 
 	return rr, err
 }
 
-func removeS3Finalizer(cr *api.PerconaXtraDBClusterBackup) {
-	filteredFins := make([]string, 0)
-
-	for _, f := range cr.GetFinalizers() {
-		if f == api.FinalizerDeleteS3Backup {
-			continue
-		}
-
-		filteredFins = append(filteredFins, f)
-	}
-
-	cr.SetFinalizers(filteredFins)
-}
-
-func (r *ReconcilePerconaXtraDBClusterBackup) tryRunS3BackupFinalizerJob(cr *api.PerconaXtraDBClusterBackup) error {
+func (r *ReconcilePerconaXtraDBClusterBackup) tryRunBackupFinalizers(ctx context.Context, cr *api.PerconaXtraDBClusterBackup) error {
 	if cr.ObjectMeta.DeletionTimestamp == nil {
 		return nil
-	}
-
-	if cr.Status.S3 == nil || cr.Status.Destination == "" || !strings.HasPrefix(cr.Status.Destination, "s3://") {
-		removeS3Finalizer(cr)
-		return nil
-	}
-
-	job := batchv1.Job{}
-	err := r.client.Get(context.TODO(), types.NamespacedName{Name: backup.GenName63(cr), Namespace: cr.Namespace}, &job)
-	if err != nil && !k8sErrors.IsNotFound(err) {
-		return errors.Wrap(err, "failed to get backup job")
-	}
-
-	if !k8sErrors.IsNotFound(err) {
-		err = r.client.Delete(context.TODO(), &job)
-		if err != nil && !k8sErrors.IsNotFound(err) {
-			return errors.Wrap(err, "failed to delete backup job")
-		}
 	}
 
 	select {
@@ -297,7 +292,7 @@ func (r *ReconcilePerconaXtraDBClusterBackup) tryRunS3BackupFinalizerJob(cr *api
 			return nil
 		}
 
-		go r.runS3BackupFinalizer(cr)
+		go r.runDeleteBackupFinalizer(ctx, cr)
 	default:
 		if _, ok := r.bcpDeleteInProgress.Load(cr.Name); !ok {
 			inprog := []string{}
@@ -314,7 +309,7 @@ func (r *ReconcilePerconaXtraDBClusterBackup) tryRunS3BackupFinalizerJob(cr *api
 	return nil
 }
 
-func (r *ReconcilePerconaXtraDBClusterBackup) runS3BackupFinalizer(cr *api.PerconaXtraDBClusterBackup) {
+func (r *ReconcilePerconaXtraDBClusterBackup) runDeleteBackupFinalizer(ctx context.Context, cr *api.PerconaXtraDBClusterBackup) {
 	logger := r.logger(cr.Name, cr.Namespace)
 
 	defer func() {
@@ -322,57 +317,124 @@ func (r *ReconcilePerconaXtraDBClusterBackup) runS3BackupFinalizer(cr *api.Perco
 		<-r.chLimit
 	}()
 
-	finalizers := []string{}
-
-	s3cli, err := r.s3cli(cr)
-	if err != nil && !k8sErrors.IsNotFound(err) {
-		logger.Error(err, "failed to create s3 client for backup", "backup", cr.Name)
-		return
-	} else if k8sErrors.IsNotFound(err) {
-		cr.SetFinalizers(finalizers)
-		return
-	}
-
+	var finalizers []string
 	for _, f := range cr.GetFinalizers() {
-		if f == api.FinalizerDeleteS3Backup {
-			logger.Info("deleting backup from s3", "name", cr.Name)
-
-			spl := strings.Split(cr.Status.Destination, "/")
-			backup := spl[len(spl)-1]
-			for _, bcp := range []string{backup + ".md5", backup + "sst_info", backup} {
-				err = retry.OnError(retry.DefaultBackoff, func(e error) bool { return true }, removeBackup(cr.Status.S3.Bucket, bcp, s3cli))
-				if err != nil {
-					break
-				}
+		var err error
+		switch f {
+		case api.FinalizerDeleteS3Backup:
+			if (cr.Status.S3 == nil && cr.Status.Azure == nil) || cr.Status.Destination == "" {
+				continue
 			}
-		} else {
+			switch cr.Status.StorageType {
+			case api.BackupStorageS3:
+				if !strings.HasPrefix(cr.Status.Destination, "s3://") {
+					continue
+				}
+				err = r.runS3BackupFinalizer(cr)
+			case api.BackupStorageAzure:
+				err = r.runAzureBackupFinalizer(ctx, cr)
+			default:
+				continue
+			}
+		default:
 			finalizers = append(finalizers, f)
 		}
+		if err != nil {
+			logger.Info("failed to delete backup", "backup path", cr.Status.Destination, "error", err.Error())
+			finalizers = append(finalizers, f)
+		} else if f == api.FinalizerDeleteS3Backup {
+			logger.Info("backup was removed", "name", cr.Name)
+		}
 	}
-
 	cr.SetFinalizers(finalizers)
 
-	if err != nil {
-		logger.Info("Failed to delete backup from s3", "backup path", cr.Status.Destination, "error", err.Error())
-	} else {
-		logger.Info("backup was removed from s3", "name", cr.Name)
-	}
-
-	err = r.client.Update(context.TODO(), cr)
+	err := r.client.Update(ctx, cr)
 	if err != nil {
 		logger.Error(err, "failed to update finalizers for backup", "backup", cr.Name)
 	}
 }
 
-func removeBackup(bucket, backup string, s3cli *minio.Client) func() error {
-	return func() error {
-		//if the bucket contains a path, we need to remove it from the bucket name and add it to the backup name
-		spl := strings.Split(bucket, "/")
-		if len(spl) > 1 {
-			bucket = spl[0]
-			backup = strings.Join(spl[1:], "/") + "/" + backup
-		}
+func (r *ReconcilePerconaXtraDBClusterBackup) runS3BackupFinalizer(cr *api.PerconaXtraDBClusterBackup) error {
+	logger := r.logger(cr.Name, cr.Namespace)
 
+	if cr.Status.S3 == nil {
+		return errors.New("s3 storage is not specified")
+	}
+
+	s3cli, err := r.s3cli(cr)
+	if err != nil && !k8sErrors.IsNotFound(err) {
+		return errors.Wrapf(err, "failed to create s3 client for backup %s", cr.Name)
+	} else if k8sErrors.IsNotFound(err) {
+		return nil
+	}
+	logger.Info("deleting backup from s3", "name", cr.Name)
+
+	spl := strings.Split(cr.Status.Destination, "/")
+	backup := spl[len(spl)-1]
+	err = retry.OnError(retry.DefaultBackoff, func(e error) bool { return true }, removeS3Backup(cr.Status.S3.Bucket, backup, s3cli))
+	if err != nil {
+		return errors.Wrapf(err, "failed to delete backup %s", cr.Name)
+	}
+	return nil
+}
+
+func (r *ReconcilePerconaXtraDBClusterBackup) runAzureBackupFinalizer(ctx context.Context, cr *api.PerconaXtraDBClusterBackup) error {
+	if cr.Status.Azure == nil {
+		return errors.New("azure storage is not specified")
+	}
+	cli, err := r.azureClient(ctx, cr)
+	if err != nil {
+		return errors.Wrap(err, "new azure client")
+	}
+	container, _ := cr.Status.Azure.ContainerAndPrefix()
+	destination := strings.TrimPrefix(cr.Status.Destination, container+"/")
+
+	err = retry.OnError(retry.DefaultBackoff,
+		func(e error) bool {
+			return true
+		},
+		removeAzureBackup(ctx, cli, container, destination))
+	if err != nil {
+		return errors.Wrapf(err, "failed to delete backup %s", cr.Name)
+	}
+	return nil
+}
+
+func removeAzureBackup(ctx context.Context, cli *azblob.Client, container, destination string) func() error {
+	return func() error {
+		blobs, err := azureListBlobs(ctx, cli, container, destination+"/")
+		if err != nil {
+			return errors.Wrap(err, "list backup blobs")
+		}
+		for _, blob := range blobs {
+			_, err = cli.DeleteBlob(ctx, container, url.QueryEscape(blob), nil)
+			if err != nil {
+				return errors.Wrapf(err, "delete blob %s", blob)
+			}
+		}
+		blobs, err = azureListBlobs(ctx, cli, container, destination+".sst_info/")
+		if err != nil {
+			return errors.Wrap(err, "list backup blobs")
+		}
+		for _, blob := range blobs {
+			_, err = cli.DeleteBlob(ctx, container, url.QueryEscape(blob), nil)
+			if err != nil {
+				return errors.Wrapf(err, "delete blob %s", blob)
+			}
+		}
+		return nil
+	}
+}
+
+func removeS3Backup(bucket, backup string, s3cli *minio.Client) func() error {
+	return func() error {
+		// this is needed to understand if user provided some path
+		// on s3, instead of just a bucket name
+		bucketSplitted := strings.Split(bucket, "/")
+		if len(bucketSplitted) > 1 {
+			bucket = bucketSplitted[0]
+			backup = strings.Join(append(bucketSplitted[1:], backup), "/")
+		}
 		objs := s3cli.ListObjects(context.Background(), bucket,
 			minio.ListObjectsOptions{
 				Recursive: true,
@@ -394,26 +456,14 @@ func removeBackup(bucket, backup string, s3cli *minio.Client) func() error {
 	}
 }
 
-func (r *ReconcilePerconaXtraDBClusterBackup) getClusterConfig(cr *api.PerconaXtraDBClusterBackup) (*api.PerconaXtraDBCluster, error) {
-	clusterList := api.PerconaXtraDBClusterList{}
-	err := r.client.List(context.TODO(),
-		&clusterList,
-		&client.ListOptions{
-			Namespace: cr.Namespace,
-		},
-	)
-
+func (r *ReconcilePerconaXtraDBClusterBackup) getCluster(cr *api.PerconaXtraDBClusterBackup) (*api.PerconaXtraDBCluster, error) {
+	cluster := api.PerconaXtraDBCluster{}
+	err := r.client.Get(context.TODO(), types.NamespacedName{Namespace: cr.Namespace, Name: cr.Spec.PXCCluster}, &cluster)
 	if err != nil {
-		return nil, errors.Wrap(err, "get clusters list")
+		return nil, errors.Wrap(err, "get PXC cluster")
 	}
 
-	for _, cluster := range clusterList.Items {
-		if cluster.Name == cr.Spec.PXCCluster {
-			return &cluster, nil
-		}
-	}
-
-	return nil, errors.Errorf("wrong cluster name: %s", cr.Spec.PXCCluster)
+	return &cluster, nil
 }
 
 func (r *ReconcilePerconaXtraDBClusterBackup) s3cli(cr *api.PerconaXtraDBClusterBackup) (*minio.Client, error) {
@@ -447,11 +497,54 @@ func (r *ReconcilePerconaXtraDBClusterBackup) s3cli(cr *api.PerconaXtraDBCluster
 		Region: cr.Status.S3.Region,
 	})
 }
+func (r *ReconcilePerconaXtraDBClusterBackup) azureClient(ctx context.Context, cr *api.PerconaXtraDBClusterBackup) (*azblob.Client, error) {
+	secret := new(corev1.Secret)
+	err := r.client.Get(ctx, types.NamespacedName{Name: cr.Status.Azure.CredentialsSecret, Namespace: cr.Namespace}, secret)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to get secret")
+	}
+	accountName := string(secret.Data["AZURE_STORAGE_ACCOUNT_NAME"])
+	accountKey := string(secret.Data["AZURE_STORAGE_ACCOUNT_KEY"])
+
+	credential, err := azblob.NewSharedKeyCredential(accountName, accountKey)
+	if err != nil {
+		return nil, errors.Wrap(err, "new credentials")
+	}
+	endpoint := fmt.Sprintf("https://%s.blob.core.windows.net/", accountName)
+	if cr.Status.Azure.Endpoint != "" {
+		endpoint = cr.Status.Azure.Endpoint
+	}
+	cli, err := azblob.NewClientWithSharedKeyCredential(endpoint, credential, nil)
+	if err != nil {
+		return nil, errors.Wrap(err, "new client")
+	}
+	return cli, nil
+}
+
+func azureListBlobs(ctx context.Context, client *azblob.Client, containerName, prefix string) ([]string, error) {
+	var blobs []string
+	pager := client.NewListBlobsFlatPager(containerName, &azblob.ListBlobsFlatOptions{
+		Prefix: &prefix,
+	})
+	for pager.More() {
+		resp, err := pager.NextPage(ctx)
+		if err != nil {
+			return nil, errors.Wrap(err, "list blobs next page")
+		}
+		if resp.Segment != nil {
+			for _, item := range resp.Segment.BlobItems {
+				if item != nil && item.Name != nil {
+					blobs = append(blobs, *item.Name)
+				}
+			}
+		}
+	}
+	return blobs, nil
+}
 
 func (r *ReconcilePerconaXtraDBClusterBackup) updateJobStatus(bcp *api.PerconaXtraDBClusterBackup, job *batchv1.Job,
-	destination, storageName string, s3 *api.BackupStorageS3Spec) error {
+	storageName string, storage *api.BackupStorageSpec, cluster *api.PerconaXtraDBCluster) error {
 	err := r.client.Get(context.TODO(), types.NamespacedName{Name: job.Name, Namespace: job.Namespace}, job)
-
 	if err != nil {
 		if k8sErrors.IsNotFound(err) {
 			return nil
@@ -461,10 +554,16 @@ func (r *ReconcilePerconaXtraDBClusterBackup) updateJobStatus(bcp *api.PerconaXt
 	}
 
 	status := api.PXCBackupStatus{
-		State:       api.BackupStarting,
-		Destination: destination,
-		StorageName: storageName,
-		S3:          s3,
+		State:                 api.BackupStarting,
+		Destination:           bcp.Status.Destination,
+		StorageName:           storageName,
+		S3:                    storage.S3,
+		Azure:                 storage.Azure,
+		StorageType:           storage.Type,
+		Image:                 bcp.Status.Image,
+		SSLSecretName:         bcp.Status.SSLSecretName,
+		SSLInternalSecretName: bcp.Status.SSLInternalSecretName,
+		VaultSecretName:       bcp.Status.VaultSecretName,
 	}
 
 	switch {
@@ -484,14 +583,22 @@ func (r *ReconcilePerconaXtraDBClusterBackup) updateJobStatus(bcp *api.PerconaXt
 
 	bcp.Status = status
 
+	if status.State == api.BackupSucceeded && cluster.PITREnabled() {
+		collectorPod, err := deployment.GetBinlogCollectorPod(context.TODO(), r.client, cluster)
+		if err != nil {
+			return errors.Wrap(err, "get binlog collector pod")
+		}
+
+		if err := deployment.RemoveGapFile(context.TODO(), r.clientcmd, collectorPod); err != nil {
+			if !errors.Is(err, deployment.GapFileNotFound) {
+				return errors.Wrap(err, "remove gap file")
+			}
+		}
+	}
+
 	err = r.client.Status().Update(context.TODO(), bcp)
 	if err != nil {
-		// may be it's k8s v1.10 and erlier (e.g. oc3.9) that doesn't support status updates
-		// so try to update whole CR
-		err := r.client.Update(context.TODO(), bcp)
-		if err != nil {
-			return errors.Wrap(err, "send update")
-		}
+		return errors.Wrap(err, "send update")
 	}
 
 	return nil
