@@ -2,6 +2,7 @@ package pxcrestore
 
 import (
 	"context"
+	"strings"
 	"time"
 
 	"github.com/pkg/errors"
@@ -15,24 +16,32 @@ import (
 	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/backup"
 )
 
-func (r *ReconcilePerconaXtraDBClusterRestore) restore(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, cluster api.PerconaXtraDBClusterSpec) error {
-	if cluster.Backup == nil {
+func (r *ReconcilePerconaXtraDBClusterRestore) restore(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, cluster *api.PerconaXtraDBCluster) error {
+	if cluster.Spec.Backup == nil {
 		return errors.New("undefined backup section in a cluster spec")
 	}
-	if len(bcp.Status.Destination) > 6 {
-		switch {
-		case bcp.Status.Destination[:4] == "pvc/":
-			return errors.Wrap(r.restorePVC(cr, bcp, bcp.Status.Destination[4:], cluster), "pvc")
-		case bcp.Status.Destination[:5] == "s3://":
-			return errors.Wrap(r.restoreS3(cr, bcp, bcp.Status.Destination[5:], cluster, false), "s3")
-		}
+	destination := bcp.Status.Destination
+	switch {
+	case strings.HasPrefix(bcp.Status.Destination, "pvc/"):
+		return errors.Wrap(r.restorePVC(cr, bcp, strings.TrimPrefix(destination, "pvc/"), cluster.Spec), "pvc")
+	case strings.HasPrefix(bcp.Status.Destination, "s3://"):
+		return errors.Wrap(r.restoreS3(cr, bcp, strings.TrimPrefix(destination, "s3://"), cluster, false), "s3")
+	case bcp.Status.Azure != nil:
+		return errors.Wrap(r.restoreAzure(cr, bcp, bcp.Status.Destination, cluster.Spec, false), "azure")
+	default:
+		return errors.Errorf("unknown backup storage type")
 	}
-
-	return errors.Errorf("unknown destination %s", bcp.Status.Destination)
 }
 
-func (r *ReconcilePerconaXtraDBClusterRestore) pitr(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, cluster api.PerconaXtraDBClusterSpec) error {
-	return errors.Wrap(r.restoreS3(cr, bcp, bcp.Status.Destination[5:], cluster, true), "PITR restore")
+func (r *ReconcilePerconaXtraDBClusterRestore) pitr(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, cluster *api.PerconaXtraDBCluster) error {
+	dest := bcp.Status.Destination
+	switch {
+	case strings.HasPrefix(dest, "s3://"):
+		return errors.Wrap(r.restoreS3(cr, bcp, strings.TrimPrefix(dest, "s3://"), cluster, true), "PITR restore s3")
+	case bcp.Status.Azure != nil:
+		return errors.Wrap(r.restoreAzure(cr, bcp, bcp.Status.Destination, cluster.Spec, true), "PITR restore azure")
+	}
+	return errors.Errorf("unknown storage type")
 }
 
 func (r *ReconcilePerconaXtraDBClusterRestore) restorePVC(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, pvcName string, cluster api.PerconaXtraDBClusterSpec) error {
@@ -81,8 +90,19 @@ func (r *ReconcilePerconaXtraDBClusterRestore) restorePVC(cr *api.PerconaXtraDBC
 
 	return r.createJob(job)
 }
+func (r *ReconcilePerconaXtraDBClusterRestore) restoreAzure(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, dest string, cluster api.PerconaXtraDBClusterSpec, pitr bool) error {
+	job, err := backup.AzureRestoreJob(cr, bcp, cluster, dest, pitr)
+	if err != nil {
+		return err
+	}
+	if err = k8s.SetControllerReference(cr, job, r.scheme); err != nil {
+		return err
+	}
 
-func (r *ReconcilePerconaXtraDBClusterRestore) restoreS3(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, s3dest string, cluster api.PerconaXtraDBClusterSpec, pitr bool) error {
+	return r.createJob(job)
+}
+
+func (r *ReconcilePerconaXtraDBClusterRestore) restoreS3(cr *api.PerconaXtraDBClusterRestore, bcp *api.PerconaXtraDBClusterBackup, s3dest string, cluster *api.PerconaXtraDBCluster, pitr bool) error {
 	job, err := backup.S3RestoreJob(cr, bcp, s3dest, cluster, pitr)
 	if err != nil {
 		return err
