@@ -1,9 +1,11 @@
 package config
 
 import (
-	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
 )
 
 func NewConfigMap(cr *api.PerconaXtraDBCluster, cmName, filename, content string) *corev1.ConfigMap {
@@ -22,23 +24,8 @@ func NewConfigMap(cr *api.PerconaXtraDBCluster, cmName, filename, content string
 	}
 }
 
-func NewAutoTuneConfigMap(cr *api.PerconaXtraDBCluster, cmName string) (*corev1.ConfigMap, error) {
-	var memory string
-
-	if cr.Spec.PXC.Resources != nil {
-		if cr.Spec.PXC.Resources.Requests != nil {
-			if len(cr.Spec.PXC.Resources.Requests.Memory) > 0 {
-				memory = cr.Spec.PXC.Resources.Requests.Memory
-			}
-		}
-		// Use limits memory in priority if it set
-		if cr.Spec.PXC.Resources.Limits != nil {
-			if len(cr.Spec.PXC.Resources.Limits.Memory) > 0 {
-				memory = cr.Spec.PXC.Resources.Limits.Memory
-			}
-		}
-	}
-	autotuneParams, err := getAutoTuneParams(memory)
+func NewAutoTuneConfigMap(cr *api.PerconaXtraDBCluster, memory *resource.Quantity, cmName string) (*corev1.ConfigMap, error) {
+	autotuneParams, err := getAutoTuneParams(cr, memory)
 	if err != nil {
 		return nil, err
 	}
