@@ -1,21 +1,26 @@
 package deployment
 
 import (
+	"bytes"
+	"context"
 	"fmt"
 	"strconv"
+	"strings"
 
+	"github.com/pkg/errors"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/percona/percona-xtradb-cluster-operator/clientcmd"
 	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
 	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/app"
-	"github.com/pkg/errors"
 )
 
 func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployment, error) {
-	storage := cr.Spec.Backup.Storages[cr.Spec.Backup.PITR.StorageName]
 	binlogCollectorName := GetBinlogCollectorDeploymentName(cr)
 	pxcUser := "xtrabackup"
 	sleepTime := fmt.Sprintf("%.2f", cr.Spec.Backup.PITR.TimeBetweenUploads)
@@ -35,23 +40,11 @@ func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployme
 	for key, value := range cr.Spec.Backup.Storages[cr.Spec.Backup.PITR.StorageName].Labels {
 		labels[key] = value
 	}
-	envs := []corev1.EnvVar{
-		{
-			Name: "SECRET_ACCESS_KEY",
-			ValueFrom: &corev1.EnvVarSource{
-				SecretKeyRef: app.SecretKeySelector(storage.S3.CredentialsSecret, "AWS_SECRET_ACCESS_KEY"),
-			},
-		},
-		{
-			Name: "ACCESS_KEY_ID",
-			ValueFrom: &corev1.EnvVarSource{
-				SecretKeyRef: app.SecretKeySelector(storage.S3.CredentialsSecret, "AWS_ACCESS_KEY_ID"),
-			},
-		},
-		{
-			Name:  "S3_BUCKET_URL",
-			Value: storage.S3.Bucket,
-		},
+	envs, err := getStorageEnvs(cr)
+	if err != nil {
+		return appsv1.Deployment{}, errors.Wrap(err, "get storage envs")
+	}
+	envs = append(envs, []corev1.EnvVar{
 		{
 			Name:  "PXC_SERVICE",
 			Value: cr.Name + "-pxc",
@@ -67,10 +60,6 @@ func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployme
 			},
 		},
 		{
-			Name:  "DEFAULT_REGION",
-			Value: cr.Spec.Backup.Storages[cr.Spec.Backup.PITR.StorageName].S3.Region,
-		},
-		{
 			Name:  "COLLECT_SPAN_SEC",
 			Value: sleepTime,
 		},
@@ -78,17 +67,7 @@ func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployme
 			Name:  "BUFFER_SIZE",
 			Value: strconv.FormatInt(bufferSize, 10),
 		},
-	}
-	if len(storage.S3.EndpointURL) > 0 {
-		envs = append(envs, corev1.EnvVar{
-			Name:  "ENDPOINT",
-			Value: storage.S3.EndpointURL,
-		})
-	}
-	res, err := app.CreateResources(cr.Spec.Backup.PITR.Resources)
-	if err != nil {
-		return appsv1.Deployment{}, errors.Wrap(err, "create resources")
-	}
+	}...)
 	container := corev1.Container{
 		Name:            "pitr",
 		Image:           cr.Spec.Backup.Image,
@@ -96,7 +75,7 @@ func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployme
 		Env:             envs,
 		SecurityContext: cr.Spec.Backup.Storages[cr.Spec.Backup.PITR.StorageName].ContainerSecurityContext,
 		Command:         []string{"pitr"},
-		Resources:       res,
+		Resources:       cr.Spec.Backup.PITR.Resources,
 		VolumeMounts: []corev1.VolumeMount{
 			{
 				Name:      "mysql-users-secret-file",
@@ -147,29 +126,145 @@ func GetBinlogCollectorDeployment(cr *api.PerconaXtraDBCluster) (appsv1.Deployme
 	}, nil
 }
 
+func getStorageEnvs(cr *api.PerconaXtraDBCluster) ([]corev1.EnvVar, error) {
+	storage := cr.Spec.Backup.Storages[cr.Spec.Backup.PITR.StorageName]
+	switch storage.Type {
+	case api.BackupStorageS3:
+		if storage.S3 == nil {
+			return nil, errors.New("s3 storage is not specified")
+		}
+		envs := []corev1.EnvVar{
+			{
+				Name: "SECRET_ACCESS_KEY",
+				ValueFrom: &corev1.EnvVarSource{
+					SecretKeyRef: app.SecretKeySelector(storage.S3.CredentialsSecret, "AWS_SECRET_ACCESS_KEY"),
+				},
+			},
+			{
+				Name: "ACCESS_KEY_ID",
+				ValueFrom: &corev1.EnvVarSource{
+					SecretKeyRef: app.SecretKeySelector(storage.S3.CredentialsSecret, "AWS_ACCESS_KEY_ID"),
+				},
+			},
+			{
+				Name:  "S3_BUCKET_URL",
+				Value: storage.S3.Bucket,
+			},
+			{
+				Name:  "DEFAULT_REGION",
+				Value: storage.S3.Region,
+			},
+			{
+				Name:  "STORAGE_TYPE",
+				Value: "s3",
+			},
+		}
+		if len(storage.S3.EndpointURL) > 0 {
+			envs = append(envs, corev1.EnvVar{
+				Name:  "ENDPOINT",
+				Value: storage.S3.EndpointURL,
+			})
+		}
+		return envs, nil
+	case api.BackupStorageAzure:
+		if storage.Azure == nil {
+			return nil, errors.New("azure storage is not specified")
+		}
+		return []corev1.EnvVar{
+			{
+				Name: "AZURE_STORAGE_ACCOUNT",
+				ValueFrom: &corev1.EnvVarSource{
+					SecretKeyRef: app.SecretKeySelector(storage.Azure.CredentialsSecret, "AZURE_STORAGE_ACCOUNT_NAME"),
+				},
+			},
+			{
+				Name: "AZURE_ACCESS_KEY",
+				ValueFrom: &corev1.EnvVarSource{
+					SecretKeyRef: app.SecretKeySelector(storage.Azure.CredentialsSecret, "AZURE_STORAGE_ACCOUNT_KEY"),
+				},
+			},
+			{
+				Name:  "AZURE_STORAGE_CLASS",
+				Value: storage.Azure.StorageClass,
+			},
+			{
+				Name:  "AZURE_CONTAINER_PATH",
+				Value: storage.Azure.ContainerPath,
+			},
+			{
+				Name:  "AZURE_ENDPOINT",
+				Value: storage.Azure.Endpoint,
+			},
+			{
+				Name:  "STORAGE_TYPE",
+				Value: "azure",
+			},
+		}, nil
+	default:
+		return nil, errors.Errorf("%s storage has unsupported type %s", cr.Spec.Backup.PITR.StorageName, storage.Type)
+	}
+}
+
 func GetBinlogCollectorDeploymentName(cr *api.PerconaXtraDBCluster) string {
 	return cr.Name + "-pitr"
 }
 
 func getBufferSize(cluster api.PerconaXtraDBClusterSpec) (mem int64, err error) {
-	var memory string
-
-	if cluster.Backup.PITR.Resources == nil {
+	res := cluster.Backup.PITR.Resources
+	if res.Size() == 0 {
 		return 0, nil
 	}
 
-	if cluster.Backup.PITR.Resources.Requests != nil && cluster.Backup.PITR.Resources.Requests.Memory != "" {
-		memory = cluster.Backup.PITR.Resources.Requests.Memory
+	var memory *resource.Quantity
+
+	if _, ok := res.Requests[corev1.ResourceMemory]; ok {
+		memory = res.Requests.Memory()
 	}
 
-	if cluster.Backup.PITR.Resources.Limits != nil && cluster.Backup.PITR.Resources.Limits.Memory != "" {
-		memory = cluster.Backup.PITR.Resources.Limits.Memory
+	if _, ok := res.Limits[corev1.ResourceMemory]; ok {
+		memory = res.Limits.Memory()
 	}
 
-	k8sQuantity, err := resource.ParseQuantity(memory)
+	return memory.Value() / int64(100) * int64(75), nil
+}
+
+func GetBinlogCollectorPod(ctx context.Context, c client.Client, cr *api.PerconaXtraDBCluster) (*corev1.Pod, error) {
+	collectorPodList := corev1.PodList{}
+
+	err := c.List(ctx, &collectorPodList,
+		&client.ListOptions{
+			Namespace: cr.Namespace,
+			LabelSelector: labels.SelectorFromSet(map[string]string{
+				"app.kubernetes.io/name":       "percona-xtradb-cluster",
+				"app.kubernetes.io/instance":   cr.Name,
+				"app.kubernetes.io/component":  "pitr",
+				"app.kubernetes.io/managed-by": "percona-xtradb-cluster-operator",
+				"app.kubernetes.io/part-of":    "percona-xtradb-cluster",
+			}),
+		},
+	)
 	if err != nil {
-		return 0, err
+		return nil, errors.Wrap(err, "get binlog collector pods")
 	}
 
-	return k8sQuantity.Value() / int64(100) * int64(75), nil
+	if len(collectorPodList.Items) < 1 {
+		return nil, errors.New("no binlog collector pods")
+	}
+
+	return &collectorPodList.Items[0], nil
+}
+
+var GapFileNotFound = errors.New("gap file not found")
+
+func RemoveGapFile(ctx context.Context, c *clientcmd.Client, pod *corev1.Pod) error {
+	stderrBuf := &bytes.Buffer{}
+	err := c.Exec(pod, "pitr", []string{"/bin/bash", "-c", "rm /tmp/gap-detected"}, nil, nil, stderrBuf, false)
+	if err != nil {
+		if strings.Contains(stderrBuf.String(), "No such file or directory") {
+			return GapFileNotFound
+		}
+		return errors.Wrapf(err, "delete gap file in collector pod %s", pod.Name)
+	}
+
+	return nil
 }
