@@ -1,14 +1,11 @@
 package statefulset
 
 import (
-	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
-	"github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/app"
-	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 )
 
-func EntrypointInitContainer(initImageName string, resources *api.PodResources, securityContext *corev1.SecurityContext, pullPolicy corev1.PullPolicy) (corev1.Container, error) {
-	c := corev1.Container{
+func EntrypointInitContainer(initImageName string, resources corev1.ResourceRequirements, securityContext *corev1.SecurityContext, pullPolicy corev1.PullPolicy) corev1.Container {
+	return corev1.Container{
 		VolumeMounts: []corev1.VolumeMount{
 			{
 				Name:      DataVolumeName,
@@ -20,12 +17,6 @@ func EntrypointInitContainer(initImageName string, resources *api.PodResources, 
 		Name:            "pxc-init",
 		Command:         []string{"/pxc-init-entrypoint.sh"},
 		SecurityContext: securityContext,
+		Resources:       resources,
 	}
-	res, err := app.CreateResources(resources)
-	if err != nil {
-		return corev1.Container{}, errors.Wrap(err, "create resources")
-	}
-	c.Resources = res
-
-	return c, nil
 }
