@@ -143,6 +143,7 @@ func (c *Node) AppContainer(spec *api.PodSpec, secrets string, cr *api.PerconaXt
 			},
 		},
 		SecurityContext: spec.ContainerSecurityContext,
+		Resources:       spec.Resources,
 	}
 
 	if cr.CompareVersionWith("1.1.0") >= 0 {
@@ -235,6 +236,12 @@ func (c *Node) AppContainer(spec *api.PodSpec, secrets string, cr *api.PerconaXt
 				SecretKeyRef: app.SecretKeySelector(secrets, "operator"),
 			},
 		})
+		if cr.CompareVersionWith("1.11.0") >= 0 && cr.Spec.PXC != nil && cr.Spec.PXC.HookScript != "" {
+			appc.VolumeMounts = append(appc.VolumeMounts, corev1.VolumeMount{
+				Name:      "hookscript",
+				MountPath: "/opt/percona/hookscript",
+			})
+		}
 	}
 
 	if cr.CompareVersionWith("1.6.0") >= 0 {
@@ -279,12 +286,6 @@ func (c *Node) AppContainer(spec *api.PodSpec, secrets string, cr *api.PerconaXt
 		appc.Env = append(appc.Env, probsEnvs...)
 	}
 
-	res, err := app.CreateResources(spec.Resources)
-	if err != nil {
-		return appc, fmt.Errorf("create resources error: %v", err)
-	}
-	appc.Resources = res
-
 	return appc, nil
 }
 
@@ -293,11 +294,6 @@ func (c *Node) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.Perc
 }
 
 func (c *Node) LogCollectorContainer(spec *api.LogCollectorSpec, logPsecrets string, logRsecrets string, cr *api.PerconaXtraDBCluster) ([]corev1.Container, error) {
-	res, err := app.CreateResources(spec.Resources)
-	if err != nil {
-		return nil, fmt.Errorf("create resources error: %v", err)
-	}
-
 	logProcEnvs := []corev1.EnvVar{
 		{
 			Name:  "LOG_DATA_DIR",
@@ -341,7 +337,7 @@ func (c *Node) LogCollectorContainer(spec *api.LogCollectorSpec, logPsecrets str
 		ImagePullPolicy: spec.ImagePullPolicy,
 		Env:             logProcEnvs,
 		SecurityContext: spec.ContainerSecurityContext,
-		Resources:       res,
+		Resources:       spec.Resources,
 		EnvFrom: []corev1.EnvFromSource{
 			{
 				SecretRef: &corev1.SecretEnvSource{
@@ -366,7 +362,7 @@ func (c *Node) LogCollectorContainer(spec *api.LogCollectorSpec, logPsecrets str
 		ImagePullPolicy: spec.ImagePullPolicy,
 		Env:             logRotEnvs,
 		SecurityContext: spec.ContainerSecurityContext,
-		Resources:       res,
+		Resources:       spec.Resources,
 		Args: []string{
 			"logrotate",
 		},
@@ -378,18 +374,27 @@ func (c *Node) LogCollectorContainer(spec *api.LogCollectorSpec, logPsecrets str
 		},
 	}
 
-	if cr.Spec.LogCollector != nil && cr.Spec.LogCollector.Configuration != "" {
-		logProcContainer.VolumeMounts = append(logProcContainer.VolumeMounts, corev1.VolumeMount{
-			Name:      "logcollector-config",
-			MountPath: "/etc/fluentbit/custom",
-		})
+	if cr.Spec.LogCollector != nil {
+		if cr.Spec.LogCollector.Configuration != "" {
+			logProcContainer.VolumeMounts = append(logProcContainer.VolumeMounts, corev1.VolumeMount{
+				Name:      "logcollector-config",
+				MountPath: "/etc/fluentbit/custom",
+			})
+		}
+
+		if cr.Spec.LogCollector.HookScript != "" {
+			logProcContainer.VolumeMounts = append(logProcContainer.VolumeMounts, corev1.VolumeMount{
+				Name:      "hookscript",
+				MountPath: "/opt/percona/hookscript",
+			})
+		}
 	}
 
 	return []corev1.Container{logProcContainer, logRotContainer}, nil
 }
 
-func (c *Node) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaXtraDBCluster) (*corev1.Container, error) {
-	ct := app.PMMClient(spec, secrets, cr.CompareVersionWith("1.2.0") >= 0, cr.CompareVersionWith("1.7.0") >= 0)
+func (c *Node) PMMContainer(spec *api.PMMSpec, secret *corev1.Secret, cr *api.PerconaXtraDBCluster) (*corev1.Container, error) {
+	ct := app.PMMClient(spec, secret, cr.CompareVersionWith("1.2.0") >= 0, cr.CompareVersionWith("1.7.0") >= 0)
 
 	pmmEnvs := []corev1.EnvVar{
 		{
@@ -403,7 +408,7 @@ func (c *Node) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaXt
 		{
 			Name: "DB_PASSWORD",
 			ValueFrom: &corev1.EnvVarSource{
-				SecretKeyRef: app.SecretKeySelector(secrets, "monitor"),
+				SecretKeyRef: app.SecretKeySelector(secret.Name, "monitor"),
 			},
 		},
 		{
@@ -430,11 +435,7 @@ func (c *Node) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaXt
 			},
 		}
 		ct.Env = append(ct.Env, clusterEnvs...)
-		res, err := app.CreateResources(spec.Resources)
-		if err != nil {
-			return nil, fmt.Errorf("create resources error: %v", err)
-		}
-		ct.Resources = res
+		ct.Resources = spec.Resources
 	}
 	if cr.CompareVersionWith("1.7.0") >= 0 {
 		for k, v := range ct.Env {
@@ -475,6 +476,23 @@ func (c *Node) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaXt
 		}
 
 	}
+	if cr.CompareVersionWith("1.10.0") >= 0 {
+		// PMM team added these flags which allows us to avoid
+		// container crash, but just restart pmm-agent till it recovers
+		// the connection.
+		sidecarEnvs := []corev1.EnvVar{
+			{
+				Name:  "PMM_AGENT_SIDECAR",
+				Value: "true",
+			},
+			{
+				Name:  "PMM_AGENT_SIDECAR_SLEEP",
+				Value: "5",
+			},
+		}
+		ct.Env = append(ct.Env, sidecarEnvs...)
+	}
+
 	ct.VolumeMounts = []corev1.VolumeMount{
 		{
 			Name:      DataVolumeName,
@@ -512,6 +530,17 @@ func (c *Node) Volumes(podSpec *api.PodSpec, cr *api.PerconaXtraDBCluster, vg ap
 		vol.Volumes = append(vol.Volumes, app.GetSecretVolumes("mysql-users-secret-file", "internal-"+cr.Name, false))
 		if cr.Spec.LogCollector != nil && cr.Spec.LogCollector.Configuration != "" {
 			vol.Volumes = append(vol.Volumes, app.GetConfigVolumes("logcollector-config", ls["app.kubernetes.io/instance"]+"-logcollector"))
+		}
+	}
+	if cr.CompareVersionWith("1.11.0") >= 0 {
+		if cr.Spec.PXC != nil && cr.Spec.PXC.HookScript != "" {
+			vol.Volumes = append(vol.Volumes,
+				app.GetConfigVolumes("hookscript", ls["app.kubernetes.io/instance"]+"-"+ls["app.kubernetes.io/component"]+"-hookscript"))
+		}
+
+		if cr.Spec.LogCollector != nil && cr.Spec.LogCollector.HookScript != "" {
+			vol.Volumes = append(vol.Volumes,
+				app.GetConfigVolumes("hookscript", ls["app.kubernetes.io/instance"]+"-logcollector-hookscript"))
 		}
 	}
 
