@@ -2,7 +2,6 @@ package statefulset
 
 import (
 	"errors"
-	"fmt"
 
 	api "github.com/percona/percona-xtradb-cluster-operator/pkg/apis/pxc/v1"
 	app "github.com/percona/percona-xtradb-cluster-operator/pkg/pxc/app"
@@ -113,6 +112,7 @@ func (c *Proxy) AppContainer(spec *api.PodSpec, secrets string, cr *api.PerconaX
 			},
 		},
 		SecurityContext: spec.ContainerSecurityContext,
+		Resources:       spec.Resources,
 	}
 	if cr.CompareVersionWith("1.9.0") >= 0 {
 		fvar := true
@@ -145,21 +145,16 @@ func (c *Proxy) AppContainer(spec *api.PodSpec, secrets string, cr *api.PerconaX
 		}
 	}
 
-	res, err := app.CreateResources(spec.Resources)
-	if err != nil {
-		return appc, fmt.Errorf("create resources error: %v", err)
+	if cr.CompareVersionWith("1.11.0") >= 0 && cr.Spec.ProxySQL != nil && cr.Spec.ProxySQL.HookScript != "" {
+		appc.VolumeMounts = append(appc.VolumeMounts, corev1.VolumeMount{
+			Name:      "hookscript",
+			MountPath: "/opt/percona/hookscript",
+		})
 	}
-	appc.Resources = res
-
 	return appc, nil
 }
 
 func (c *Proxy) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.PerconaXtraDBCluster) ([]corev1.Container, error) {
-	res, err := app.CreateResources(spec.SidecarResources)
-	if err != nil {
-		return nil, fmt.Errorf("create sidecar resources error: %v", err)
-	}
-
 	pxcMonit := corev1.Container{
 		Name:            "pxc-monit",
 		Image:           spec.Image,
@@ -169,7 +164,7 @@ func (c *Proxy) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.Per
 			"-on-change=/usr/bin/add_pxc_nodes.sh",
 			"-service=$(PXC_SERVICE)",
 		},
-		Resources: res,
+		Resources: spec.SidecarResources,
 		Env: []corev1.EnvVar{
 			{
 				Name:  "PXC_SERVICE",
@@ -209,7 +204,7 @@ func (c *Proxy) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.Per
 			"-on-change=/usr/bin/add_proxysql_nodes.sh",
 			"-service=$(PROXYSQL_SERVICE)",
 		},
-		Resources: res,
+		Resources: spec.SidecarResources,
 		Env: []corev1.EnvVar{
 			{
 				Name:  "PROXYSQL_SERVICE",
@@ -239,6 +234,16 @@ func (c *Proxy) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.Per
 			},
 		},
 	}
+	if cr.Spec.AllowUnsafeConfig && (cr.Spec.TLS == nil || cr.Spec.TLS.IssuerConf == nil) {
+		pxcMonit.Env = append(pxcMonit.Env, corev1.EnvVar{
+			Name:  "SSL_DIR",
+			Value: "/dev/null",
+		})
+		proxysqlMonit.Env = append(proxysqlMonit.Env, corev1.EnvVar{
+			Name:  "SSL_DIR",
+			Value: "/dev/null",
+		})
+	}
 	if cr.CompareVersionWith("1.9.0") >= 0 {
 		fvar := true
 		envFrom := corev1.EnvFromSource{
@@ -266,12 +271,12 @@ func (c *Proxy) SidecarContainers(spec *api.PodSpec, secrets string, cr *api.Per
 	return []corev1.Container{pxcMonit, proxysqlMonit}, nil
 }
 
-func (c *Proxy) LogCollectorContainer(spec *api.LogCollectorSpec, logPsecrets string, logRsecrets string, cr *api.PerconaXtraDBCluster) ([]corev1.Container, error) {
+func (c *Proxy) LogCollectorContainer(_ *api.LogCollectorSpec, _ string, _ string, _ *api.PerconaXtraDBCluster) ([]corev1.Container, error) {
 	return nil, nil
 }
 
-func (c *Proxy) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaXtraDBCluster) (*corev1.Container, error) {
-	ct := app.PMMClient(spec, secrets, cr.CompareVersionWith("1.2.0") >= 0, cr.CompareVersionWith("1.7.0") >= 0)
+func (c *Proxy) PMMContainer(spec *api.PMMSpec, secret *corev1.Secret, cr *api.PerconaXtraDBCluster) (*corev1.Container, error) {
+	ct := app.PMMClient(spec, secret, cr.CompareVersionWith("1.2.0") >= 0, cr.CompareVersionWith("1.7.0") >= 0)
 
 	pmmEnvs := []corev1.EnvVar{
 		{
@@ -285,7 +290,7 @@ func (c *Proxy) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaX
 		{
 			Name: "MONITOR_PASSWORD",
 			ValueFrom: &corev1.EnvVarSource{
-				SecretKeyRef: app.SecretKeySelector(secrets, "monitor"),
+				SecretKeyRef: app.SecretKeySelector(secret.Name, "monitor"),
 			},
 		},
 	}
@@ -298,7 +303,7 @@ func (c *Proxy) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaX
 		{
 			Name: "DB_PASSWORD",
 			ValueFrom: &corev1.EnvVarSource{
-				SecretKeyRef: app.SecretKeySelector(secrets, "monitor"),
+				SecretKeyRef: app.SecretKeySelector(secret.Name, "monitor"),
 			},
 		},
 		{
@@ -325,11 +330,7 @@ func (c *Proxy) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaX
 	ct.Env = append(ct.Env, pmmEnvs...)
 	if cr.CompareVersionWith("1.2.0") >= 0 {
 		ct.Env = append(ct.Env, dbEnvs...)
-		res, err := app.CreateResources(spec.Resources)
-		if err != nil {
-			return nil, fmt.Errorf("create resources error: %v", err)
-		}
-		ct.Resources = res
+		ct.Resources = spec.Resources
 	} else {
 		ct.Env = append(ct.Env, dbArgsEnv...)
 	}
@@ -368,6 +369,23 @@ func (c *Proxy) PMMContainer(spec *api.PMMSpec, secrets string, cr *api.PerconaX
 		ct.Env = append(ct.Env, pmmAgentScriptEnv...)
 	}
 
+	if cr.CompareVersionWith("1.10.0") >= 0 {
+		// PMM team added these flags which allows us to avoid
+		// container crash, but just restart pmm-agent till it recovers
+		// the connection.
+		sidecarEnvs := []corev1.EnvVar{
+			{
+				Name:  "PMM_AGENT_SIDECAR",
+				Value: "true",
+			},
+			{
+				Name:  "PMM_AGENT_SIDECAR_SLEEP",
+				Value: "5",
+			},
+		}
+		ct.Env = append(ct.Env, sidecarEnvs...)
+	}
+
 	return &ct, nil
 }
 
@@ -388,7 +406,10 @@ func (c *Proxy) Volumes(podSpec *api.PodSpec, cr *api.PerconaXtraDBCluster, vg a
 	if err == nil {
 		vol.Volumes = append(vol.Volumes, configVolume)
 	}
-
+	if cr.CompareVersionWith("1.11.0") >= 0 && cr.Spec.ProxySQL != nil && cr.Spec.ProxySQL.HookScript != "" {
+		vol.Volumes = append(vol.Volumes,
+			app.GetConfigVolumes("hookscript", ls["app.kubernetes.io/instance"]+"-"+ls["app.kubernetes.io/component"]+"-hookscript"))
+	}
 	return vol, nil
 }
 
